hacky hack: exclude C++ compilers older than v4.0.0
We have to exclude LLVM 3.8 because of its early implementation of OpenMP that seems to be different
from the library it inherits from GCC 6.3.0 . Everything seems to compile starting from LLVM 4.0

However we don't have an adequate language yet to express that only LLVM should be 4.0.0 or younger.
Given that other compilers available to me at the moment have higher versions, I'm opting for this
unfair comparison of apples to oranges.